+++ src/App.tsx
@@ -6,7 +6,6 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Partner from "./pages/Partner";
 import Legal from "./pages/Legal";
 import NotFound from "./pages/NotFound";
-import ConsentBanner from "./components/landing/ConsentBanner";
 
 const queryClient = new QueryClient();
 
@@ -24,7 +23,6 @@ const App = () => (
           {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
           <Route path="*" element={<NotFound />} />
         </Routes>
-        <ConsentBanner />
       </BrowserRouter>
     </TooltipProvider>
   </QueryClientProvider>